gas 32-bit host compile warnings
	* config/tc-d10v.c (find_opcode): Correct format specifier vs.
	arg mismatch.
	* config/tc-m68hc11.c (fixup8, fixup16, fixup24, fixup8_xg): Likewise.
	* config/tc-vax.c (md_assemble): Likewise.
	* config/tc-xtensa.c (dump_litpools): Likewise.
	* config/tc-z80.c (emit_data_val, emit_byte): Likewise.

+++ gas/config/tc-xtensa.c
@@ -7869,8 +7869,9 @@ dump_litpools (void)
 		count++;
 	      litfrag = litfrag->fr_next;
 	    }
-	  printf("   %ld <%d:%d> (%d) [%d]: ",
-		 lpf->addr, lpf->priority, lpf->original_priority,
+	  printf("   %" PRId64 " <%d:%d> (%d) [%d]: ",
+		 (int64_t) lpf->addr,
+		 lpf->priority, lpf->original_priority,
 		 lpf->fragP->fr_line, count);
 	  /* dump_frag(lpf->fragP);  */
 	}
